Major refactor: renamed packages and updated build structure
There's been a large amount of trial and error over the past few patches as I try to figure out what works and what doesn't. This particular change was for several reasons.

1) There was some confusion over what the purpose of the json files were. The package was renamed to "source_data" to be more clear.

2) All python packages were moved to mhwdata. Apparently its a common recommendation to name your "main package" after the project. Lets see how it works.

3) Renamed some packages to reduce ambiguity. src.data is now mhwdata.io. build.db is now mhwdb.sql. Loading data was moved from /data to mhwdata.load.

The data loading and build routines will be refactored at a later date.
