+++ mhwdata/sql/mappings.py
