+++ source_data/armors/armor_base.json
