+++ source_data/armors/armor_data.json
