+++ source_data/armors/armorset_base.json
