+++ source_data/decorations/decoration_base.json
