+++ source_data/monsters/monster_base.json
