+++ source_data/monsters/monster_rewards.json
