+++ source_data/monsters/monster_weaknesses.json
