+++ source_data/weapons/weapon_base.json
