+++ source_data/weapons/weapon_data/bow.json
