+++ source_data/weapons/weapon_data/charge-blade.json
