+++ source_data/weapons/weapon_data/great-sword.json
