+++ source_data/weapons/weapon_data/hammer.json
