+++ source_data/weapons/weapon_data/heavy-bowgun.json
