+++ source_data/weapons/weapon_data/insect-glaive.json
