+++ source_data/weapons/weapon_data/light-bowgun.json
