+++ source_data/weapons/weapon_data/long-sword.json
