+++ source_data/weapons/weapon_data/switch-axe.json
