+++ source_data/weapons/weapon_data/sword-and-shield.json
